+++ Kits/rokt/rokt/Sources/mParticle-Rokt/MPKitRokt.m
@@ -3,7 +3,7 @@
 @import RoktContracts;
 
 // Kit version
-static NSString * const kMPRoktKitVersion = @"9.0.1";
+static NSString * const kMPRoktKitVersion = @"9.1.0";
 
 // Constants for kit configuration keys
 static NSString * const kMPKitConfigurationIdKey = @"id";